backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
sqlalchemy
psycopg2-binary